#124 chore: 마이카이빙 패키지명 변경
