+++ android/MyCarchiving/app/src/main/java/com/softeer/mycarchiving/ui/HyundaiAppState.kt
@@ -30,9 +30,9 @@ import com.softeer.mycarchiving.ui.makingcar.navigateToMakingCar
 import com.softeer.mycarchiving.ui.makingcar.selectcolor.navigateToSelectColor
 import com.softeer.mycarchiving.ui.makingcar.selectoption.navigateToSelectOption
 import com.softeer.mycarchiving.ui.makingcar.selecttrim.navigateToSelectTrim
-import com.softeer.mycarchiving.ui.myarchive.myarchivedetail.navigateToMyArchiveDetail
-import com.softeer.mycarchiving.ui.myarchive.myarchivemade.navigateToMyArchiveMade
-import com.softeer.mycarchiving.ui.myarchive.myarchivesave.navigateToMyArchiveSave
+import com.softeer.mycarchiving.ui.myarchive.detail.navigateToMyArchiveDetail
+import com.softeer.mycarchiving.ui.myarchive.made.navigateToMyArchiveMade
+import com.softeer.mycarchiving.ui.myarchive.save.navigateToMyArchiveSave
 import com.softeer.mycarchiving.ui.myarchive.navigateToMyArchiving
 import com.softeer.mycarchiving.util.MakeCarProcess
 import kotlinx.coroutines.CoroutineScope
